MiKoCodeFixProvider introduced as base class for all CodeFixProviders
